Move rivet version validate-config to rivet config validate
